правки по замечаниям

+++ Task__6/src/main/java/ru/otus/homeWork/shellservice/AuthorService.java
@@ -7,6 +7,9 @@ import org.springframework.shell.standard.ShellMethod;
 import org.springframework.transaction.annotation.Transactional;
 import ru.otus.homeWork.domain.Author;
 import ru.otus.homeWork.repositories.AuthorRepositoryJpa;
+import ru.otus.homeWork.uiservice.AuthorUIService;
+
+import java.util.List;
 
 
 @ShellComponent
@@ -16,43 +19,42 @@ public class AuthorService {
 
     private final AuthorRepositoryJpa authorRepo;
 
-    @Transactional(readOnly = true)
+    private final AuthorUIService uiService;
+
     @ShellMethod(value = "Authors count", key = "a_count")
     public void authorsCount() {
-        System.out.println(authorRepo.getAuthorsCount());
+        uiService.print(authorRepo.getAuthorsCount());
     }
 
     @Transactional
     @ShellMethod(value = "Save new author", key = "a_save")
-    public void save(String firstName, String lastName) {
-        Author author = new Author();
-        author.setFirstName(firstName);
-        author.setLastName(lastName);
-        authorRepo.save(author);
-        System.out.println("new author is saved");
+    public void save() {
+        Author author = uiService.save();
+        boolean isSaved = authorRepo.save(author) != null;
+        uiService.isSaved(isSaved);
     }
 
-    @Transactional(readOnly = true)
     @ShellMethod(value = "Get author by Id number", key = "a_get")
-    public void getById(long id) {
-        Author author = authorRepo.getById(id);
-        if(author == null)
-            System.out.println("Wrong id. Record is missing");
-        else
-            System.out.println(author);
+    public void getById() {
+        Long id = uiService.getId();
+        if(id != null) {
+            Author author = authorRepo.getById(id);
+            uiService.print(author);
+        }
     }
 
-    @Transactional(readOnly = true)
     @ShellMethod(value = "Get all authors", key = "a_getAll")
     public void getAll() {
-        authorRepo.getAll().stream().forEach(System.out::println);
+        List<Author> authors = authorRepo.getAll();
+        uiService.print(authors);
     }
-
 
     @Transactional
     @ShellMethod(value = "Delete author by Id number", key = "a_delete")
-    public void deleteById(long id) {
-        authorRepo.deleteById(id);
-        System.out.println("author is deleted");
+    public void deleteById() {
+        Long id = uiService.getId();
+        if(id != null){
+            authorRepo.deleteById(id);
+        }
     }
 }
