+++ Task__6/src/main/java/ru/otus/homeWork/shellservice/CommentService.java
@@ -9,6 +9,8 @@ import ru.otus.homeWork.domain.Book;
 import ru.otus.homeWork.domain.Comment;
 import ru.otus.homeWork.repositories.BookRepository;
 import ru.otus.homeWork.repositories.CommentRepository;
+import ru.otus.homeWork.uiservice.BookUIService;
+import ru.otus.homeWork.uiservice.CommentUIService;
 
 @ShellComponent
 @ShellCommandGroup(value = "CRUD for comment")
@@ -19,34 +21,57 @@ public class CommentService {
 
     private final BookRepository bookRepo;
 
+    private final CommentUIService commentUIService;
+
+    private final BookUIService bookUIService;
+
     @Transactional
     @ShellMethod(value = "Save new comment", key = "c_save")
-    public void save(String content, Long  book_id) {
-        Book book = bookRepo.getById(book_id);
-        if(book == null) {
-            System.out.printf("Book with id=%d is not exists\n", book_id);
+    public void save() {
+        boolean isSaved = false;
+        Comment comment = commentUIService.save();
+        if(comment == null) {
+            commentUIService.isSaved(false);
             return;
         }
-        Comment comment = new Comment();
-        comment.setContent(content);
-        comment.setBook(book);
-        commentRepo.save(comment);
+
+        Long bookId = comment.getBook().getId();
+        Book book = bookRepo.getById(bookId);
+        commentUIService.isExists(book, bookId);
+
+        if(book != null) {
+            comment.setBook(book);
+            commentRepo.save(comment);
+            isSaved = true;
+        }
+        commentUIService.isSaved(isSaved);
     }
 
-    @Transactional(readOnly = true)
     @ShellMethod(value = "Get comment by Id number", key = "c_get")
-    public void getById(long id) {
-        Comment comment = commentRepo.getById(id);
-        if(comment == null)
-            System.out.println("Wrong id. Record is missing");
-        else
-            System.out.println(comment);
+    public void getById() {
+        Long id = commentUIService.getId();
+        if(id != null) {
+            Comment comment = commentRepo.getById(id);
+            commentUIService.print(comment);
+        }
     }
 
     @Transactional
     @ShellMethod(value = "Delete comment by Id number", key = "c_delete")
-    public void deleteById(long id) {
-        commentRepo.deleteById(id);
-        System.out.println("The comment is deleted");
+    public void deleteById() {
+        Long id = commentUIService.getId();
+        if(id != null) {
+            commentRepo.deleteById(id);
+        }
+    }
+
+    @Transactional(readOnly = true)
+    @ShellMethod(value = "Get book comments", key = "c_getBookComments")
+    public void getAllComments() {
+        Long id = commentUIService.getId();
+        if(id != null) {
+            Book book = bookRepo.getById(id);
+            commentUIService.print(book.getComments());
+        }
     }
 }
