+++ Task__6/src/main/java/ru/otus/homeWork/shellservice/GenreService.java
@@ -7,6 +7,10 @@ import org.springframework.shell.standard.ShellMethod;
 import org.springframework.transaction.annotation.Transactional;
 import ru.otus.homeWork.domain.Genre;
 import ru.otus.homeWork.repositories.GenreRepository;
+import ru.otus.homeWork.uiservice.GenreUIService;
+
+import java.util.List;
+
 @ShellComponent
 @ShellCommandGroup(value = "CRUD for genre")
 @RequiredArgsConstructor
@@ -14,41 +18,42 @@ public class GenreService {
 
     private final GenreRepository genreRepo;
 
-    @Transactional(readOnly = true)
+    private final GenreUIService uiService;
+
     @ShellMethod(value = "Genres count", key = "g_count")
     public void getGenresCount() {
-        System.out.println(genreRepo.getGenresCount());
+        uiService.print(genreRepo.getGenresCount());
     }
 
     @Transactional
     @ShellMethod(value = "Save new genre", key = "g_save")
-    public void save(String name) {
-        Genre genre = new Genre();
-        genre.setName(name);
-        genreRepo.save(genre);
-        System.out.println("new genre is saved");
+    public void save() {
+        Genre genre = uiService.save();
+        boolean isSaved = genreRepo.save(genre) != null;
+        uiService.isSaved(isSaved);
     }
 
-    @Transactional(readOnly = true)
     @ShellMethod(value = "Get genre by Id number", key = "g_get")
-    public void getById(long id) {
-        Genre genre = genreRepo.getById(id);
-        if(genre == null)
-            System.out.println("Wrong id. Record is missing");
-        else
-            System.out.println(genre);
+    public void getById() {
+        Long id = uiService.getId();
+        if(id != null) {
+            Genre genre = genreRepo.getById(id);
+            uiService.print(genre);
+        }
     }
 
-    @Transactional(readOnly = true)
     @ShellMethod(value = "Get all genres", key = "g_getAll")
     public void getAll() {
-        genreRepo.getAll().stream().forEach(System.out::println);
+        List<Genre> genres = genreRepo.getAll();
+        uiService.print(genres);
     }
 
     @Transactional
     @ShellMethod(value = "Delete genre by Id number", key = "g_delete")
-    public void deleteById(long id) {
-        genreRepo.deleteById(id);
-        System.out.println("The genre is deleted");
+    public void deleteById() {
+        Long id = uiService.getId();
+        if(id != null) {
+            genreRepo.deleteById(id);
+        }
     }
 }
